PyMaker/datas/object3/2nd/i501/Python_DOC_referPages.compare-expected-vs-actual-json-response-for-a-get-request-in-python.txt

Natural Text
You will get a small json response back when you go to this site https://reqres.in/api/users/2I am saving the response in a variable(actual). I have also put the response in another variable(expected).Both responses are same. I am changing the values to test failed cases. The ultimate goal is to compare 2 and make sure they match.I have 2 functions, 1 compares keys and value of both dictionaries and the other function sorts the dictionaries.  Code below:Problem: The test passes when there is no difference.However if there is any difference the order goes crazy. Here is an example where I changed data to dat inside expected response:Expected data  but got dat  Expected id  but got last_name  Expected 2  but got WeaverShould the sort function be more specific rather than using sort_keys=True?By the way thought about **args but I don't think that is a good choice in this scenario.Thank You for your expert comment and time.
I advise to use unittest and avoid using so much nested for loopsAt this point, the will get you this :Now, you can easily make test in a class. 
Key order is not guaranteed in Python versions under 3.7; you should use collections.OrderedDict when you need to create an object that remembers key order.In Python 3.7 the insertion ordered is preserved, so your keys will always match.


Answer URL
https://docs.python.org/3/library/collections.html#collections.OrderedDict
